fix(types): resolve mypy errors gating #353 Type Check job
13 pre-existing annotation mismatches surfaced once Lint passed and Type
Check could run:
- strategy_spec: narrow validation_plan/cost_model/required_gates via temp vars
- sub_gates edge/monthly: declare metrics as dict[str, float | None]
- replay_parity: max(key=lambda) instead of overloaded dict.get
- audit_cli: (gb or {}).get on union-optional gate row
- taker_engine: guard positions is not None before len/index
- trade_pnl_projector: widen price_scale to float (floated internally)

+++ research/backtest/trade_pnl_projector.py
@@ -124,7 +124,7 @@ def project_trade_pnl_from_position_series(
 def project_trade_pnl(
     fills: list[dict[str, Any]],
     *,
-    price_scale: int = _DEFAULT_PRICE_SCALE,
+    price_scale: float = _DEFAULT_PRICE_SCALE,
 ) -> list[float]:
     """Return per-round-trip PnL in points via FIFO matching.
